+++ src/Functions/Add-VNet-integration-to-Function-App.ps1
@@ -3,7 +3,8 @@ param (
     [Parameter(Mandatory)][string] $FunctionAppResourceGroupName,
     [Parameter(Mandatory)][string] $FunctionAppName,
     [Alias("VnetName")]
-    [Parameter(Mandatory)][string] $FunctionAppVnetIntegrationName,
+    [Alias("FunctionAppVnetIntegrationName")]
+    [Parameter(Mandatory)][string] $FunctionAppVnetIntegrationVnetName,
     [Parameter(Mandatory)][string] $FunctionAppVnetIntegrationSubnetName,
     [Parameter()][string] $FunctionAppServiceDeploymentSlotName,
     [Parameter()][bool] $ApplyToAllSlots = $false
@@ -22,7 +23,8 @@ function Add-VnetIntegration
         [Parameter(Mandatory)][string] $FunctionAppResourceGroupName,
         [Parameter(Mandatory)][string] $FunctionAppName,
         [Alias("VnetName")]
-        [Parameter(Mandatory)][string] $FunctionAppVnetIntegrationName,
+        [Alias("FunctionAppVnetIntegrationName")]
+        [Parameter(Mandatory)][string] $FunctionAppVnetIntegrationVnetName,
         [Parameter(Mandatory)][string] $FunctionAppVnetIntegrationSubnetName,
         [Parameter()][string] $FunctionAppServiceDeploymentSlotName
     )
@@ -32,21 +34,22 @@ function Add-VnetIntegration
     $fullFunctionAppName = $FunctionAppName
     $additionalParameters = @()
 
-    if ($FunctionAppServiceDeploymentSlotName) {
+    if ($FunctionAppServiceDeploymentSlotName)
+    {
         $additionalParameters += '--slot' , $FunctionAppServiceDeploymentSlotName
         $fullFunctionAppName += " [$FunctionAppServiceDeploymentSlotName]"
     }
 
     $vnetIntegrations = Invoke-Executable az functionapp vnet-integration list --resource-group $FunctionAppResourceGroupName --name $FunctionAppName @additionalParameters | ConvertFrom-Json
-    $matchedIntegrations = $vnetIntegrations | Where-Object  vnetResourceId -like "*/providers/Microsoft.Network/virtualNetworks/$FunctionAppVnetIntegrationName/subnets/$FunctionAppVnetIntegrationSubnetName"
-    if($matchedIntegrations)
+    $matchedIntegrations = $vnetIntegrations | Where-Object  vnetResourceId -like "*/providers/Microsoft.Network/virtualNetworks/$FunctionAppVnetIntegrationVnetName/subnets/$FunctionAppVnetIntegrationSubnetName"
+    if ($matchedIntegrations)
     {
         Write-Host "VNET Integration found for $fullFunctionAppName"
     }
     else
     {
         Write-Host "VNET Integration not found, adding it to $fullFunctionAppName"
-        Invoke-Executable az functionapp vnet-integration add --resource-group $FunctionAppResourceGroupName --name $FunctionAppName --vnet $FunctionAppVnetIntegrationName --subnet $FunctionAppVnetIntegrationSubnetName @additionalParameters
+        Invoke-Executable az functionapp vnet-integration add --resource-group $FunctionAppResourceGroupName --name $FunctionAppName --vnet $FunctionAppVnetIntegrationVnetName --subnet $FunctionAppVnetIntegrationSubnetName @additionalParameters
         Invoke-Executable az functionapp restart --name $FunctionAppName --resource-group $FunctionAppResourceGroupName
     }
 
@@ -63,12 +66,12 @@ if ($ApplyToAllSlots)
 }
 
 # Set VNET Integration on the main given slot (normally production)
-Add-VnetIntegration -FunctionAppResourceGroupName $FunctionAppResourceGroupName -FunctionAppName $FunctionAppName -FunctionAppVnetIntegrationName $FunctionAppVnetIntegrationName -FunctionAppVnetIntegrationSubnetName $FunctionAppVnetIntegrationSubnetName -FunctionAppServiceDeploymentSlotName $FunctionAppServiceDeploymentSlotName
+Add-VnetIntegration -FunctionAppResourceGroupName $FunctionAppResourceGroupName -FunctionAppName $FunctionAppName -FunctionAppVnetIntegrationVnetName $FunctionAppVnetIntegrationVnetName -FunctionAppVnetIntegrationSubnetName $FunctionAppVnetIntegrationSubnetName -FunctionAppServiceDeploymentSlotName $FunctionAppServiceDeploymentSlotName
 
 # Apply to all slots if desired
-foreach($availableSlot in $availableSlots)
+foreach ($availableSlot in $availableSlots)
 {
-    Add-VnetIntegration -FunctionAppResourceGroupName $FunctionAppResourceGroupName -FunctionAppName $FunctionAppName -FunctionAppVnetIntegrationName $FunctionAppVnetIntegrationName -FunctionAppVnetIntegrationSubnetName $FunctionAppVnetIntegrationSubnetName -FunctionAppServiceDeploymentSlotName $availableSlot.name
+    Add-VnetIntegration -FunctionAppResourceGroupName $FunctionAppResourceGroupName -FunctionAppName $FunctionAppName -FunctionAppVnetIntegrationVnetName $FunctionAppVnetIntegrationVnetName -FunctionAppVnetIntegrationSubnetName $FunctionAppVnetIntegrationSubnetName -FunctionAppServiceDeploymentSlotName $availableSlot.name
 }
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet