+++ src/Log-Analytics-Workspace/Get-Log-Analytics-Workspace-Id-for-Pipeline.ps1
@@ -1,7 +1,8 @@
 [CmdletBinding()]
 param (
     [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceResourceGroupName,
-    [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceName
+    [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceName,
+    [Parameter()][string] $OutputPipelineVariableName = "LogAnalyticsWorkspaceId"
 )
 
 #region ===BEGIN IMPORTS===
@@ -10,8 +11,8 @@ Import-Module "$PSScriptRoot\..\AzDocs.Common" -Force
 
 Write-Header -ScopedPSCmdlet $PSCmdlet
 
-$logAnalyticsWorkspaceResourceId = (az monitor log-analytics workspace show --resource-group $LogAnalyticsWorkspaceResourceGroupName --workspace-name $LogAnalyticsWorkspaceName | ConvertFrom-Json).id
-Write-Host "LogAnalyticsWorkspaceResourceId: $LogAnalyticsWorkspaceResourceId"
-Write-Host "##vso[task.setvariable variable=LogAnalyticsWorkspaceResourceId]$LogAnalyticsWorkspaceResourceId"
+$logAnalyticsWorkspaceId = (Invoke-Executable az monitor log-analytics workspace show --resource-group $LogAnalyticsWorkspaceResourceGroupName --workspace-name $LogAnalyticsWorkspaceName | ConvertFrom-Json).customerId
+Write-Host "LogAnalyticsWorkspaceId: $logAnalyticsWorkspaceId"
+Write-Host "##vso[task.setvariable variable=$($OutputPipelineVariableName);isOutput=true]$logAnalyticsWorkspaceId"
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet